move font def to layout
